+++ ZapLibUnitTest/RegExpTests.cs
@@ -1,11 +1,8 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using ZapLib;
+using Newtonsoft.Json;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Diagnostics;
+using System.Text.RegularExpressions;
 
 namespace ZapLib.Tests
 {
@@ -24,14 +21,30 @@ namespace ZapLib.Tests
 
 
             RegExp reg = new RegExp(@"Name: (\w+) Age: (\d+)");
-            string[] result =  reg.Exec(contnet);
+            string[] result = reg.Exec(contnet);
 
-            foreach(string s in result)
+            foreach (string s in result)
             {
                 Console.WriteLine(s);
             }
 
             Assert.IsNotNull(result);
+        }
+
+        [TestMethod()]
+        public void ExecTest()
+        {
+
+            RegExp reg = new RegExp(@"[A-Z]123456789",RegexOptions.None);
+
+            Trace.WriteLine("A123456789");
+            string[] res1 = reg.Exec("A123456789");
+            Trace.WriteLine(JsonConvert.SerializeObject(res1));
+
+            Trace.WriteLine("a123456789");
+            string[] res2 = reg.Exec("a123456789");
+            Trace.WriteLine(JsonConvert.SerializeObject(res2));
+
         }
     }
 }